+++ hc/accounts/urls.py
@@ -13,7 +13,7 @@ urlpatterns = [
     path("signup/", views.signup, name="hc-signup"),
     path("login_link_sent/", views.login_link_sent, name="hc-login-link-sent"),
     path(
-        "check_token/<slug:username>/<str:token>/",
+        "check_token/<str:username>/<str:token>/",
         views.check_token,
         name="hc-check-token",
     ),